:wrench: Add config to catch static error
