+++ v3/gatsby-theme-baitong/package.json
@@ -12,6 +12,9 @@
   "dependencies": {
     "@emotion/core": "^10.0.16",
     "@emotion/styled": "^10.0.15",
+    "@fortawesome/fontawesome-svg-core": "^1.2.25",
+    "@fortawesome/free-solid-svg-icons": "^5.11.2",
+    "@fortawesome/react-fontawesome": "^0.1.6",
     "@mdx-js/react": "^1.4.0",
     "gatsby-image": "^2.2.16",
     "gatsby-plugin-emotion": "^4.1.4",
